Fix port number scanning
Lots of issues with port number scanning.  IPv6 addresses wouldn't
work, and address families couldn't be specified.

So redo the scanning of the port to scan for a valid state string, then
use that to determing where the port number ends.

Allow the port number to begin with "ipv4," or "ipv6," to specify a
protocol family.

Signed-off-by: Corey Minyard <cminyard@mvista.com>