Remove commas as an option separator
This was only documented in one place and never demonstrated and a
really bad idea.  It prevented commas inside of options as separators,
which prevented network addresses from being specified in options.
Plus, commas would be very hard to read as option separators.

It is, unfortunately, a behavior change, but I don't see a way around
it.  I apologize in advance if this affects anyone.

Signed-off-by: Corey Minyard <cminyard@mvista.com>