Fix several bugs with UDP and remote address additions
The address length on sendto was not zero on TCP, this can result
in issues for some systems.

kickolduser mode was completely broken.  It got a lot more complicated
because we have to actually receive the packet/accept now, so we have to
store off the remote information (and the packet in UDP mode) to allow
the old user to be kicked off.

If a use was kicked before devstr or banner data was transmitted, it
might not free the data.  Free the old data before adding new data.

Signed-off-by: Corey Minyard <cminyard@mvista.com>