Fix a race between fixed remaddr and kickolduser
If you have a fixed address added on a reconfig on a netcon that
was just kicked, it's possible to get an fixed remaddr set up on
a port that was going to be started up with a new user.  So just
kick the new user off in this case.

Signed-off-by: Corey Minyard <cminyard@mvista.com>