genio: Finish getting SSL working
Add a IN_CHILD_OPEN and IN_CHILD_CLOSE to the state machine to know
because in some cases it's necessary to know if we are opening the
child or we are doing the SSL negotiation, or doing the SSL shutdown
or the child close.

Move all the SSL_write processing to the write callback handler, doing
it in the ssln_write() routine was redundant.

Allocate two times the read size for the SSL BIO, because it needs more
space than just the full read size.

If the child reports a read error, go straight to SSLN_IN_CHILD_CLOSE
on a close and don't try to do SSL shutdown, because the child is
in a bad state.

Use a signed value for the return of BIO_write do comparison with -1
works.

Don't call SSL_write unless some bytes are ready to process.  Calling
it with a zero size makes is return zero, which is an error.

Fix ssln_set_child_enables() to cover all situations and don't set them
anyplace else.

Make some functions private that were not but needed to be.

Signed-off-by: Corey Minyard <cminyard@mvista.com>