Fix shutdown handling to write all data and be timed
Modify the shutdown sequence to make sure that all pending data is
written to the port before closing it.  Also make sure that the
shutdown finishes by adding timeout handling so that if the data
is not written in a reasonable amount of time to go ahead and shut
it down.

Signed-off-by: Corey Minyard <cminyard@mvista.com>